sorting amd searching added

+++ src/components/axios.js
@@ -17,7 +17,7 @@ instance.interceptors.request.use((req) => {
 })
 
 instance.interceptors.response.use((response) => response, (error) => {
-    if (error.response.status == 401) {
+    if (error && error.response && error.response.status == 401) {
         localStorage.clear()
         window.location = '/login';
     }
